refactor: add preprocessor naming compatibility aliases
