+++ experimental/preprocessor/__init__.py
@@ -7,9 +7,10 @@ from .constants import (
     PRECOMPILER_NO_DIRECTIVE_SENTINEL,
     PrecompileOutcome,
 )
-from .heuristic_precompiler import PrecompileResult, precompile_heuristic
+from .heuristic_preprocessor import PrecompileResult, precompile_heuristic
 from .output_validation import (
     parse_precompiler_output,
+    parse_preprocessor_output,
     validate_precompiler_output,
 )
 from .prompt_utils import render_prompt
@@ -21,6 +22,7 @@ __all__ = [
     "PRECOMPILER_NO_DIRECTIVE_SENTINEL",
     "PrecompileResult",
     "PrecompileOutcome",
+    "parse_preprocessor_output",
     "parse_precompiler_output",
     "precompile_heuristic",
     "render_prompt",
